Update testsuite; call_indirect has new syntax (#667)
The `call_indirect` instruction now allows the type to be specified
inline or via the common "type use" syntax (or both):

```
call_indirect (type $t1) ...
call_indirect (param i32 f32) (result f64) ...
call_indirect (type $t2) (param i32) ...
```

This means that `CallIndirectExpr` changes from storing a `Var` (the
referenced func type) to a `FuncDeclaration`, which can store both a
type use and a function signature. Most of the changes here are fallout
from that change.

The other major change is that function signature resolution and error
checking now needs to iterate over a functions expr list looking for
`call_indirect` instructions. I'm not sure if this is a significant
overhead in parsing/validation yet, but there are plenty of ways to
optimize it if it ends up in profiles.