Use Streams in the WriteBinarySpecScript API (#744)
`WriteBinarySpecScript` writes a JSON file and its associated module
files (either `.wasm` or `.wat` files). The previous API wrote directly
to the filesystem, but it is more useful to allow the caller to specify
where to write the output instead.

The function now provides a callback `WriteBinarySpecStreamFactory`,
which takes a filename as parameter and returns a `Stream` to write to.

There is also a convenience function that will write the results to a
vector of filename/MemoryStream pairs.