Add support for v8x16.swizzle and the load_splats. (#1116)
* Add support for v8x16.shuffle1 and v8x16.shuffle2_imm.

v8x16.shuffle2_imm is a rename of the previous v8x16.shuffle, but I add a copy of the code as if it were a new instruction in case the spec proposal makes further changes. The tests for old v8x16.shuffle remain in place and while there are new tests for the new v8x16.shuffle1, there are not for v8x16.shuffle2_imm. The behaviour and implementation are the same as for v8x16.shuffle, so we should simply search and replace the existing tests at some point, leaving one of them untested, probably the deprecated v8x16.shuffle.

I did test v8x16.shuffle1 against the SIMD spec test from WAVM and it passes. The WAVM spec tests for v8x16.shuffle2_imm parse but it has no invocations of the instruction.

* Rename v8x16.shuffle1 and x8v16.shuffle2_imm to v8x16.swizzle and v8x16.shuffle_imm.

* Update SIMD operands.

 * Swizzle is just a binary operator.
 * Shuffle is named "v8x16.shuffle".
 * Add 4 new opcodes for load_splat.
 * Remove legacy 0xfd 0x03 opcode for shuffle.

* Test all four load splats.