Switch python indentation from 2-space to 4-space (#1145)
pep8 specifies 4 space indentation.  The use of 2 spaces is, I believe,
a historical anomaly where certain large organizations such as google
chose 2 over 4 and have yet to make the switch.

For a project like wabt with little python code I think the cost of
switching is small enough to justify the churn.