Check for MSVC first in list of compiler tests (#1209)
This allows use of clang-cl.exe, wherin Clang pretends to be MSVC and accepts its flags.
It will be treated as MSVC in this case.

There are a few places where we'll want to tune further, and perhaps refactor to allow a compiler to be both MSVC and clang, or perhaps use clang's flags instead of MSVC's. But this gets things working.