Avoid os.path.relpath in gen-spec-js.py (#1246)
relpath here was failing with on windows when TMPDIR is on a different
drive.  This is because `new_module_filename` lives in the TMPDIR
where and json_dir might be on different drive.

These paths don't get embedded in the final JS anyway because we
embed them directly binary strings so its not important that they are
relative.