Always check index before accessing vectors (#1347)
Since the validator tries to print as many errors as possible, it
shouldn't stop when the index is invalid (since this may produce future
errors). Instead it uses a default value.

This is encoded in a new function called CheckIndexWithValue, which
either accesses the vector, or uses a default value.