Fix interpreter garbage collector (#1357)
The interpreter has a garbage colletor (Store::Collect), that *almost*
worked.

In particular, the free list had no fast way to determine whether an
object was used or not. This is necessary when iterating over objects to
actually delete them. This commit adds an `is_free_` vector for this
purpose.

The `Store::Collect` method was also not correct; it was calling
`Store::Mark(Ref)` on each of the GC roots, but that only marks the ref
itself, it doesn't iterate over the Object's references. This commit
updates Collect to iterate over all marked objects iteratively until
steady-state is reached. It's potentially quadratic in the number of
objects as currently written, so there's definitely room for
improvement.