Eager symbol tables (#1537)
* Eagerly create symbol tables when writing relocatable binaries

Instead of adding entries to the symbol table as they are referenced,
when writing relocatable binaries, we're going to make symbols for all
functions.  This allows exported functions and globals to be written
with the proper exported / no_strip flags, so that resulting files will
link with wasm-ld.

* Symbol table names are export names, without the dollar.

In the previous commit, I wrongly assumed that the globally visible name
for linking was taken from the exports section, whereas actually it's
from the symbol table.  Therefore this patch changes to strip off the
dollar, and also to make all named bindings globally visible.  The
exported-to-the-host binding is mostly unrelated to the
visible-to-other-compilation-units binding.  Unnamed definitions aren't
added to the symbol table.

* Add symbol tables test

* Rename Intern helper to EnsureUnique