Add clang-format check to CI (#1683)
This adds clang-format check to the CI. This only checks the diff so it
doesn't affect the other parts of the codebase. Also generated files in
src/prebuilt/ directories are excluded from the check. A new
src/prebuilt/.clang-format is added to disable the check in the
directory.

scripts/clang-format-diff.sh is copied from the same file we are using in
Binaryen.