[simd] Correctly shift alignment_log2 (#1699)
The alignment value in binary format is log2, so we need to shift it in
binary-reader-ir before it is validated (since validation requires that
it is shifted to be the number of bytes). We correctly did that for some
Simd instructions (like load splat) but did not do it for load/store
lane.

Fixes #1674.