Enable reference types by default (#1729)
This features was finished earlier this year:
https://github.com/WebAssembly/proposals/blob/master/finished-proposals.md

One thing to note is that the version of the spec tests we currently
have in third_party/testsuite doesn't have ref types merged yet so
this change disables ref types when running some of those tests.  This
can be removed in a followup when we update the testsuite.