Add support for inline export/import of tags (#1747)
This PR adds support for inline export/import of tags in modules. e.g.,

```
(tag (import "foo" "bar") (param i32))
```

This is used in the spec tests for the exception handling proposal already:

  https://github.com/WebAssembly/exception-handling/blob/main/test/core/tag.wast

so it would be useful to support in wabt, so that the test suite can be imported in the future.