Fix function body start/end locations in wasm-validate (#1842)
For text validation, this means the error is always correctly reported on
the final expression in the function.

For binary validation, this means that we report the byte after the last
instruction in the function as the failure location.  This is in line
with other binary validation reports.  For example, for `type mismatch
in i32.add` we report the validation error at the byte *after* the add
instruction.