wast2json: write binary modules verbatim (#1932)
Previously any top-level module defined in a .wast file would be parsed
and converted to a `Module` and stored in a `ModuleCommand`, since it
may be used later by the script (e.g. via an `invoke` command). This
means that when it was written out later it may have been changed, which
is undesirable for binary and quoted modules.

This change adds a new `ScriptModuleCommand`, which stores both a
`Module` and a `ScriptModule`. The `Module` contains the parsed
information, and the `ScriptModule` contains the original contents of
the binary or quoted data.

This fixes issue #1927.