Fix git describe behavior (#1979)
* Rename CMAKE_PROJECT_VERSION to WABT_VERSION_STRING

CMAKE_PROJECT_VERSION is a standard variable that is controlled by the
project() command. Renaming our usage to WABT_VERSION_STRING better
describes the purpose of the variable and avoids surprises down the
line.

* Use git describe output to annotate version string

Before this commit, the attempt to add the output of git describe to the
version string printed by $wasm-tool --version was bugged. It would
always print the three-part version number (like 1.0.29).

After this commit, the output of `git describe --tags` is used to append
a description to the version number in the form of: "ver (git~desc)".
For example:

    $ ./wasm2c --version
    1.0.29 (git~1.0.29-27-gf63184ef)

If this command returns a tag equal in name to the current version,
nothing is appended.

If this command fails, a NOTICE (rather than a WARNING) is printed (only
if the project is top-level), and only the bare version number is used.
If the source code is not living in the git repository then, again, only
the bare number is used, but no warning is printed.

Fixes #1977
Fixes #1978