Fix architecture checks (#1969)
* Use standard modules to test endianness

CMake prior to v3.20 provides a module TestBigEndian which we can use to
set the WABT_BIG_ENDIAN define. As of 3.20, the CMAKE_<LANG>_BYTE_ORDER
variable should be preferred. Leaving this as a note for the future.

* Fix x87 math detection

TARGET_ARCH was only used to determine whether to add gcc-specific SSE
math flags (-msse2 -mfpmath=sse). The new approach simply assumes gcc
compatibility with its __i386__ and __SSE2_MATH__ symbols, which are
defined precisely when we are targeting x86-32 with SSE2 math enabled.
If those macros are defined, then we conclude all is well. Otherwise, we
add the flags if we know the compiler is gcc or clang (and will thus
accept them) and issue a warning if the compiler is unknown.

Fixes #1709
Fixes #1688