libwabt.js: Add `check` option, equiv to wasm2wat's --no-check (#2582)
## Description

When using the JS API, there's currently no way to get the equivalent of
wasm2wat's `--no-check` option. This adds support for that.

- Add a property named `check` to the`options` arg of`readWasm` (in
wasm.post.js)
- Add a checkbox to the demo to control the `check` option.

## Screenshots

Here's the wasm2wat demo showing the new option, and with an example of
wasm2wat on an invalid module:

![CleanShot 2025-04-09 at 13 35
57@2x](https://github.com/user-attachments/assets/40077a89-7676-4934-8323-a10c32001f08)

(Note that the function body doesn't end with `end`.)