Support table initializer expressions; completes function references support (#2593)
Another large patch on the top of the other two. Now there are 3-4000
lines of new code is waiting for review, and it looks like the end is
still far.

The painful part so far:

https://webassembly.github.io/function-references/core/binary/modules.html#element-section

Since type 0-3 is reserved for `(ref func)`, all `(elem ... funcref
...)` must be encoded with 4-7, which increase binary size. This affect
several tests, and probably real word programs :(

There are spec tests to check this behavior.